pyproject.toml
[project]
name = "bruce"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.3",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "kafka-python>=2.0.2",
    "psycopg2-binary>=2.9.10",
]
